requirements.txt
jira>=2.0.0
google-cloud-secret-manager>=2.0.0